requirements_cronjob.txt
# Cronjob Server Dependencies
apscheduler==3.10.4
httpx==0.27.0
supabase==2.3.4
python-dotenv==1.0.0
